WIP!
switch to react-native-tab-navigator.  it renders more efficiently since it
only renders the active tab, yet allows all of the tabbed content to be mounted

its currently pinned to my branch, but hoping my upstream PR gets accepted.

styling is a little tricky, so if we go with this then we'll need to figure out
and document a good default and make sure its easy to customize